Refactor Github dates and API code into separate files
