+++ follow_top_repos_by_star_count.py
@@ -1,33 +1,11 @@
 from typing import List
-
-from github import Github
-
 from lgtm import LGTMSite
-from datetime import datetime
+
+import github_dates
+import github_api
 import sys
 import yaml
 import time
-
-def create_github() -> Github:
-    with open("config.yml") as config_file:
-        config = yaml.safe_load(config_file)
-        github: dict = config['github']
-        return Github(github['api_key'])
-
-def current_year() -> int:
-    now = datetime.now()
-    return now.year
-
-def generate_dates() -> List[str]:
-    date_ranges: List[str] = []
-
-    # Github was formed in 2008
-    year_range = list(range(2008, current_year() + 1))
-
-    for i, year in enumerate(year_range):
-        date_ranges.append(f'{year}-01-01..{year}-12-31')
-
-    return date_ranges
 
 def save_project_to_lgtm(site: 'LGTMSite', repo_name: str):
     print("Adding: " + repo_name)
@@ -41,10 +19,10 @@ def save_project_to_lgtm(site: 'LGTMSite', repo_name: str):
     print("Saved the project: " + repo_name)
 
 def find_and_save_projects_to_lgtm(language: str):
-    github = create_github()
+    github = github_api.create()
     site = LGTMSite.create_from_file()
 
-    for date_range in generate_dates():
+    for date_range in github_dates.generate_dates():
         repos = github.search_repositories(query=f'stars:>500 created:{date_range} sort:stars language:{language}')
 
         for repo in repos:
