+++ github_api.py
@@ -1,0 +1,7 @@
+import yaml
+
+def create() -> Github:
+    with open("config.yml") as config_file:
+        config = yaml.safe_load(config_file)
+        github: dict = config['github']
+        return Github(github['api_key'])
