+++ test.py
@@ -1,0 +1,3 @@
+import github_dates
+
+print(github_dates.current_year())
